+++ OpenWeight.Cli/Commands/DaemonCommand.cs
@@ -1,0 +1,12 @@
+﻿using Spectre.Console.Cli;
+using Aijkl.OpenWeight.Settings;
+
+namespace OpenWeight.Commands;
+
+internal class DaemonCommand : AsyncCommand<DaemonSettings>
+{
+    public override Task<int> ExecuteAsync(CommandContext context, DaemonSettings settings)
+    {
+        throw new NotImplementedException();
+    }
+}
