+++ OpenWeight.Cli/OpenWeight.Cli.csproj
@@ -13,8 +13,4 @@
     <PackageReference Include="Spectre.Console.Cli" Version="0.46.0" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Settings\" />
-  </ItemGroup>
-
 </Project>
