+++ OpenWeight.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.33209.295
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenWeight", "OpenWeight\OpenWeight.csproj", "{CF680F53-4AFF-424D-BE27-84A0E664E167}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenWeight.Cli", "OpenWeight.Cli\OpenWeight.Cli.csproj", "{CF680F53-4AFF-424D-BE27-84A0E664E167}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
