--- OpenWeight/Commands/DaemonCommand.cs
@@ -1,12 +1,0 @@
-﻿using Spectre.Console.Cli;
-
-namespace OpenWeight.Commands
-{
-    internal class DaemonCommand : AsyncCommand
-    {
-        public override Task<int> ExecuteAsync(CommandContext context)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
